add feature manage category for admin

+++ internal/service/category_usecases.go
@@ -1,25 +1,43 @@
 package service
 
 import (
-	"errors"
-
 	"forum/internal/entities"
 	"forum/internal/repository"
+	"forum/pkg/validator"
 )
 
 type CategoryUseCase struct {
 	categoryRepo repository.CategoryRepository
+}
+type CategoryForm struct {
+	Name string
+	validator.Validator
+}
+
+func (uc *CategoryUseCase) NewCategoryCreateForm() CategoryForm {
+	return CategoryForm{}
 }
 
 func NewCategoryUseCase(categoryRepo repository.CategoryRepository) *CategoryUseCase {
 	return &CategoryUseCase{categoryRepo: categoryRepo}
 }
 
-func (u *CategoryUseCase) Insert(name string) (int, error) {
-	if name == "" {
-		return 0, errors.New("category name can't be empty")
+func (u *CategoryUseCase) Insert(form *CategoryForm) (int, error) {
+	form.CheckField(validator.NotBlank(form.Name), "category", "This field cannot be blank")
+	form.CheckField(validator.Matches(form.Name, validator.TextRX), "category", "This field must contain only english or russian letters")
+	exists, err := u.categoryRepo.ExistName(form.Name)
+	if err != nil {
+		return 0, err
 	}
-	return u.categoryRepo.Insert(name)
+	if exists {
+		form.CheckField(false, "category", "error dublicate name")
+
+	}
+	if !form.Valid() {
+		return 0, entities.ErrInvalidData
+	}
+
+	return u.categoryRepo.Insert(form.Name)
 }
 
 func (u *CategoryUseCase) Get(categoryId int) (*entities.Category, error) {
@@ -38,5 +56,9 @@ func (u *CategoryUseCase) Delete(categoryId int) error {
 	if !exists {
 		return entities.ErrNoRecord
 	}
+	if categoryId == 1 {
+		return entities.ErrInvalidData
+	}
+
 	return u.categoryRepo.Delete(categoryId)
 }
